requirements.txt
biopython>=1.79
pytest>=6.2.5
matplotlib
parameterized
python-levenshtein